rename unsigned int to size_t

See merge request TSML/TSC!9

+++ include/cry/cbc.h
@@ -21,7 +21,7 @@ struct cry_cbc_ctx {
     /** Counter */
     unsigned char              ctr[CRY_CBC_BLOCK_MAX];
     /** Counter length */
-    unsigned int               ctrlen;
+    size_t                     ctrlen;
 };
 
 typedef struct cry_cbc_ctx cry_cbc_ctx;
@@ -48,7 +48,7 @@ void cry_cbc_init(struct cry_cbc_ctx *ctx, void *ciph_ctx,
  * @param size  Cipher key size.
  */
 void cry_cbc_key_set(struct cry_cbc_ctx *ctx, const unsigned char *key,
-                     unsigned int size);
+                     size_t size);
 
 /**
  * Set the initialization vector.
@@ -58,7 +58,7 @@ void cry_cbc_key_set(struct cry_cbc_ctx *ctx, const unsigned char *key,
  * @param size  Initialization vector size.
  */
 void cry_cbc_iv_set(struct cry_cbc_ctx *ctx, const unsigned char *iv,
-                    unsigned int size);
+                    size_t size);
 
 /**
  * Encryption function.
@@ -69,7 +69,7 @@ void cry_cbc_iv_set(struct cry_cbc_ctx *ctx, const unsigned char *iv,
  * @param size  Size should be a multiple of CBC_BLOCK_SIZE
  */
 void cry_cbc_encrypt(struct cry_cbc_ctx *ctx, unsigned char *dst,
-                     const unsigned char *src, unsigned int size);
+                     const unsigned char *src, size_t size);
 
 /**
  * Decryption function.
@@ -80,7 +80,7 @@ void cry_cbc_encrypt(struct cry_cbc_ctx *ctx, unsigned char *dst,
  * @param size  Size should be a multiple of CBC_BLOCK_SIZE.
  */
 void cry_cbc_decrypt(struct cry_cbc_ctx *ctx, unsigned char *dst,
-                     const unsigned char *src, unsigned int size);
+                     const unsigned char *src, size_t size);
 
 #ifdef __cplusplus
 }
