+++ include/cry/des.h
@@ -14,7 +14,7 @@
 
 /** DES context */
 struct cry_des_ctx {
-    unsigned int  keylen; /* 8 for normal DES, 24 for triple DES */
+    size_t        keylen; /* 8 for normal DES, 24 for triple DES */
     unsigned char key[3*CRY_DES_BLOCK_SIZE];
 };
 
@@ -22,12 +22,12 @@ typedef struct cry_des_ctx cry_des_ctx;
 
 
 void cry_des_key_set(cry_des_ctx *ctx, unsigned char *key,
-                     unsigned int size);
+                     size_t size);
 
 void cry_des_encrypt(cry_des_ctx *ctx, unsigned char *dst,
-                     const unsigned char *src, unsigned int size);
+                     const unsigned char *src, size_t size);
 
 void cry_des_decrypt(cry_des_ctx *ctx, unsigned char *dst,
-                     const unsigned char *src, unsigned int size);
+                     const unsigned char *src, size_t size);
 
 #endif /* CRY_DES_H_ */
