+++ test/hmac_test.c
@@ -3,9 +3,9 @@
 #include <cry/sha256.h>
 
 struct hmac_param {
-    unsigned int  keylen;
-    unsigned int  msglen;
-    unsigned int  maclen;
+    size_t keylen;
+    size_t msglen;
+    size_t maclen;
     unsigned char key[256];
     unsigned char msg[256];
     unsigned char mac[64];
